Change badge fk from article to event for simplicity
